Update the guide to use Springboot 3.1.3 (#136)
* Update to use Spring Boot 3.1.3 and springBoot-3.0 feature

* Update workflow to use Java 17 for tests

* use Java 17

* update the spring-boot app

* update the spring-boot app

* update the output

* update the output

* update the output

* check Java version

* enable Java 17

* enable Java 17

* enable Java 17

* enable Java 17

* enable Java 17

* update daily build test scripts

* update daily build test scripts

* update daily build test scripts

* run docker logs if test fails

* Update testApp.sh

* Update testApp.sh

* Update dailyBuild.sh

* Update pom.xml

* Update dailyBuild.sh

* Update dailyBuild.sh

* Update README.adoc

---------
