+++ README.adoc
@@ -10,6 +10,7 @@
 :projectid: spring-boot
 :page-duration: 20 minutes
 :page-releasedate: 2019-11-04
+:page-majorupdateddate: 2023-11-15
 :page-description: Learn how to containerize, package, and run a Spring Boot application on Open Liberty.
 :page-tags: ['Docker']
 :page-related-guides: ['rest-intro', 'containerize']
@@ -28,7 +29,7 @@ Learn how to containerize, package, and run a Spring Boot application on Open Li
 
 == What you'll learn
 
-The starting point of this guide is the finished application from the https://spring.io/guides/gs/spring-boot/[Building an Application with Spring Boot^] guide. If you are not familiar with Spring Boot, complete that guide first. Java 8 is required to run this project.
+The starting point of this guide is the finished application from the https://spring.io/guides/gs/spring-boot/[Building an Application with Spring Boot^] guide. If you are not familiar with Spring Boot, complete that guide first. Java 17 is required to run this project.
 
 You will learn how to use the `springBootUtility` command to deploy a Spring Boot application in Docker on Open Liberty without modification. This command stores the dependent library JAR files of the application to the target library cache, and packages the remaining application artifacts into a thin application JAR file.
 
@@ -69,10 +70,15 @@ You can now run the application in the embedded Tomcat web container by executin
 java -jar target/guide-spring-boot-0.1.0.jar
 ```
 
-Notice that the console output displays a message about the application running in Tomcat on port `8080`. 
+After you see the following messages, the application is ready:
 [role='no_copy']
 ----
-... INFO ... [ main] o.s.b.w.embedded.tomcat.TomcatWebServer : Tomcat started on port(s): 8080 (http) with context path ''
+... INFO ... [ main] com.example.springboot.Application : Started Application in 2.511 seconds (process running for 3.24)
+Let's inspect the beans provided by Spring Boot:
+application
+...
+welcomePageHandlerMapping
+welcomePageNotAcceptableHandlerMapping
 ----
 
 // Static guide instruction
