--- finish/src/main/java/hello/HelloController.java
@@ -1,14 +1,0 @@
-package hello;
-
-import org.springframework.web.bind.annotation.RestController;
-import org.springframework.web.bind.annotation.RequestMapping;
-
-@RestController
-public class HelloController {
-    
-    @RequestMapping("/hello")
-    public String hello() {
-        return "Greetings from Spring Boot!";
-    }
-    
-}
