+++ finish/src/main/liberty/config/server.xml
@@ -3,10 +3,10 @@
 
     <featureManager>
     <!-- tag::servlet[] -->
-        <feature>servlet-4.0</feature>
+        <feature>servlet-6.0</feature>
     <!-- end::servlet[] -->
     <!-- tag::springboot[] -->
-        <feature>springBoot-2.0</feature>
+        <feature>springBoot-3.0</feature>
     <!-- end::springboot[] -->
     </featureManager>
