+++ scripts/dockerImageTest.sh
@@ -9,13 +9,14 @@ done
 
 echo "Testing latest OpenLiberty Docker image"
 
-sed -i "\#<artifactId>liberty-maven-plugin</artifactId>#,\#<configuration>#c<artifactId>liberty-maven-plugin</artifactId><version>3.2.3</version><configuration><install><runtimeUrl>https://public.dhe.ibm.com/ibmdl/export/pub/software/openliberty/runtime/nightly/""$DATE""/""$DRIVER""</runtimeUrl></install>" pom.xml
+sed -i "\#<artifactId>liberty-maven-plugin</artifactId>#,\#<configuration>#c<artifactId>liberty-maven-plugin</artifactId><version>3.8.2</version><configuration><install><runtimeUrl>https://public.dhe.ibm.com/ibmdl/export/pub/software/openliberty/runtime/nightly/""$DATE""/""$DRIVER""</runtimeUrl></install>" pom.xml
 cat pom.xml
 
-sed -i "s;FROM icr.io/appcafe/open-liberty:kernel-slim-java8-openj9-ubi;FROM openliberty/daily:latest;g" Dockerfile
+sed -i "s;FROM icr.io/appcafe/open-liberty:full-java17-openj9-ubi;FROM openliberty/daily:latest;g" Dockerfile
+sed -i "s;FROM icr.io/appcafe/open-liberty:kernel-slim-java17-openj9-ubi;FROM openliberty/daily:latest;g" Dockerfile
 sed -i "s;RUN features.sh;#RUN features.sh;g" Dockerfile
 cat Dockerfile
 
 docker pull -q "openliberty/daily:latest"
 
-sudo ../scripts/testApp.sh
+sudo -E ../scripts/testApp.sh
