+++ start/src/main/java/com/example/springboot/HelloController.java
@@ -1,0 +1,14 @@
+package com.example.springboot;
+
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+public class HelloController {
+
+	@GetMapping("/hello")
+	public String index() {
+		return "Greetings from Spring Boot!";
+	}
+
+}
